[setup] cleaned code
