+++ src/main/java/actions/exec/Run.java
@@ -1,0 +1,10 @@
+
+package actions.exec;
+
+public class Run {
+
+    public static void main(String[] args) {
+
+    }
+
+}
