+++ src/main/java/actions/metier/Action.java
@@ -1,0 +1,5 @@
+package actions.metier;
+
+public abstract class Action {
+
+}
