+++ src/main/java/actions/metier/ActionComposee.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+public class ActionComposee extends Action {
+
+}
