+++ src/main/java/actions/metier/ActionSimple.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+public class ActionSimple extends Action {
+
+}
