+++ src/main/java/actions/metier/Catalogue.java
@@ -1,0 +1,5 @@
+package actions.metier;
+
+public class Catalogue {
+
+}
