+++ src/main/java/actions/metier/Cours.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+public class Cours {
+
+}
