+++ src/main/java/actions/metier/Portefeuille.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+public class Portefeuille {
+
+}
