--- src/main/java/tp04/exec/Run.java
@@ -1,73 +1,0 @@
-/*
- * Copyright 2025 David Navarre &lt;David.Navarre at irit.fr&gt;.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tp04.exec;
-
-import tp04.metier.Action;
-import tp04.metier.ActionComposee;
-import tp04.metier.ActionSimple;
-import tp04.metier.Jour;
-import tp04.metier.Portefeuille;
-
-public class Run {
-
-    public static void main(String[] args) {
-        ActionSimple bnp, axa;
-        ActionComposee bqAss;
-        Jour j1, j2;
-
-        // init des objets metiers Jour
-        j1 = new Jour(2014, 1);
-        j2 = new Jour(2014, 2);
-
-        // creation d'actions simples et composée
-        bnp = new ActionSimple("BNP");
-        axa = new ActionSimple("AXA");
-        bqAss = new ActionComposee("Banque-Assurance");
-        // enrg de la composition de l'action composée
-        bqAss.enrgComposition(axa, 0.3f);
-        bqAss.enrgComposition(bnp, 0.7f);
-        // enrg. de 2 cours pour chaque action 
-        axa.enrgCours(j1, 200);
-        axa.enrgCours(j2, 250);
-        bnp.enrgCours(j1, 100);
-        bnp.enrgCours(j2, 200);
-        // affichage des cours - comme 1 action simple et 1 action
-        System.out.println("Action simple *bnp* à j1 : " + bnp.valeur(j1));
-        System.out.println("Action *Banque-Assurance* à j2 : " + bqAss.valeur(j2));
-
-        Portefeuille p;
-        p = new Portefeuille();
-        p.acheter(axa, 10);
-        System.out.println("Portefeuille : " + p);
-        p.acheter(bnp, 20);
-        System.out.println("Portefeuille : " + p);
-        p.acheter(bqAss, 5);
-        System.out.println("Portefeuille : " + p);
-        p.acheter(bqAss, 15);
-        System.out.println("Portefeuille : " + p);
-        System.out.println("Portefeuille à j1 : " + p.valeur(j1));
-        p.vendre(axa, 5);
-        System.out.println("Portefeuille : " + p);
-        p.vendre(axa, 5);
-        System.out.println("Portefeuille : " + p);
-        p.vendre(axa, 5);
-        System.out.println("Portefeuille : " + p);
-        p.vendre(bnp, 50);
-        System.out.println("Portefeuille : " + p);
- 
-    }
-
-}
