--- src/main/java/tp04/metier/Action.java
@@ -1,68 +1,0 @@
-/*
- * Copyright 2025 David Navarre &lt;David.Navarre at irit.fr&gt;.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tp04.metier;
-
-import java.util.Objects;
-
-/**
- *
- * @author perussel
- */
-public abstract class Action {
-
-    private String libelle;
-
-    /**
-     * Get the value of libelle
-     *
-     * @return the value of libelle
-     */
-    public String getLibelle() {
-        return libelle;
-    }
-
-    public Action(String libelle) {
-        this.libelle = libelle;
-    }
-
-    public abstract float valeur(Jour j);
-
-    @Override
-    public int hashCode() {
-        int hash = 3;
-        hash = 53 * hash + Objects.hashCode(this.libelle);
-        return hash;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (obj == null) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        final Action other = (Action) obj;
-        if (!Objects.equals(this.libelle, other.libelle)) {
-            return false;
-        }
-        return true;
-    }
-
-    public String toString() {
-        return this.getLibelle();
-    }
-}
