--- src/main/java/tp04/metier/ActionComposee.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2025 David Navarre &lt;David.Navarre at irit.fr&gt;.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tp04.metier;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- *
- * @author perussel
- */
-public class ActionComposee extends Action {
-    // attribut lien
-    Map<ActionSimple, Float> mapPanier;
-
-    public ActionComposee(String libelle) {
-        super(libelle);
-        this.mapPanier = new HashMap();
-    }
-
-    public void enrgComposition(ActionSimple as, float pourcentage) {
-        this.mapPanier.put(as, pourcentage);
-    }
-
-    @Override
-    public float valeur(Jour j) {
-        float valeur;
-
-        valeur = 0;
-        for (ActionSimple as : this.mapPanier.keySet()) {
-            valeur = valeur + (as.valeur(j) * this.mapPanier.get(as));
-        }
-
-        return valeur;
-    }
-
-}
