--- src/main/java/tp04/metier/ActionSimple.java
@@ -1,73 +1,0 @@
-/*
- * Copyright 2025 David Navarre &lt;David.Navarre at irit.fr&gt;.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tp04.metier;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- *
- * @author perussel
- */
-public class ActionSimple extends Action {
-
-    // attribut lien
-    private Map<Jour, Cours> mapCours;
-
-    // constructeur
-    public ActionSimple(String libelle) {
-        // Action simple initialisée comme 1 action
-        super(libelle);
-        // init spécifique
-        this.mapCours = new HashMap();
-    }
-
-    // enrg possible si pas de cours pour ce jour
-    public void enrgCours(Jour j, float v) {
-        if (this.mapCours.containsKey(j) == false)
-            this.mapCours.put(j, new Cours(j, v));
-    }
-
-    @Override
-    public float valeur(Jour j) {
-        if (this.mapCours.containsKey(j) == true)
-            return this.mapCours.get(j).getValeur();
-        else
-            return 0; // definition d'une constante possible
-    }
-
-    // encapsulation de la définition de la classe Cours
-    private class Cours {
-
-        private Jour jour;
-
-        private float valeur;
-
-        public float getValeur() {
-            return valeur;
-        }
-
-        public Jour getJour() {
-            return jour;
-        }
-
-        public Cours(Jour jour, float valeur) {
-            this.jour = jour;
-            this.valeur = valeur;
-        }
-
-    }
-}
