+++ src/test/java/actions/metier/ActionComposeeTest.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+public class ActionComposeeTest {
+
+}
