+++ src/test/java/actions/metier/ActionTest.java
@@ -1,0 +1,6 @@
+
+package actions.metier;
+
+class ActionTest {
+
+}
