--- src/test/java/tp04/metier/ActionTest.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2025 David Navarre &lt;David.Navarre at irit.fr&gt;.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package tp04.metier;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-/**
- *
- * @author David Navarre &lt;David.Navarre at irit.fr&gt;
- */
-class ActionTest {
-
-    @Test
-    void testGetLibelle() {
-        final Action action = new ActionImpl();
-        Assertions.assertNotNull(action.getLibelle());
-    }
-
-    public class ActionImpl extends Action {
-
-        public ActionImpl() {
-            super("");
-        }
-
-        public float valeur(Jour j) {
-            return 0.0F;
-        }
-    }
-
-}
